added new route for update customer, Created single component of form for update and create customer, Removed Unnecessary code.
